Fixed a new bunch of incorrect field types in database loading
Former-commit-id: bd61f023af19d05c422242d6ef2cda81efaf1ba5

+++ src/game/AuctionHouseMgr.cpp
@@ -342,7 +342,7 @@ void AuctionHouseMgr::LoadAuctions()
     }
 
     Field *fields = result->Fetch();
-    uint32 AuctionCount=fields[0].GetUInt32();
+    uint32 AuctionCount=fields[0].GetUInt64();
 
     if(!AuctionCount)
     {
